Refactor some variable names and use macro for creating external links
