+++ src/main/app/forms/validation/validationErrors.ts
@@ -11,7 +11,7 @@ export class ValidationErrors {
   static readonly NON_NEGATIVE_NUMBER_REQUIRED: string = 'Don’t enter a negative number'
   static readonly POSITIVE_NUMBER_REQUIRED: string = 'Enter a number higher than 0'
   static readonly INTEGER_REQUIRED: string = 'Enter a numeric, for example 3'
-  static readonly OVER_18_DEFENDANT_REQUIRED: string = 'Select yes, no, or company/organisation'
+  static readonly DEFENDANT_AGE_REQUIRED: string = 'Select yes, no, or company/organisation'
 
   static readonly DATE_REQUIRED: string = 'Enter a date'
   static readonly DATE_INVALID_YEAR: string = 'Enter a 4 digit year'
