+++ src/main/features/claim/routes/eligibility/defendant-age.ts
@@ -12,20 +12,20 @@ import { NotEligibleReason } from 'claim/helpers/eligibility/notEligibleReason'
 import { ValidationGroups } from 'claim/helpers/eligibility/validationGroups'
 import { Draft } from '@hmcts/draft-store-client'
 import { DraftClaim } from 'drafts/models/draftClaim'
-import { DefendantAgeOption } from 'claim/form/models/eligibility/DefendantAgeOption'
+import { DefendantAgeOption } from 'claim/form/models/eligibility/defendantAgeOption'
 
 function renderView (form: Form<Eligibility>, res: express.Response): void {
-  res.render(Paths.eligibilityDefendantAgeOptionPage.associatedView, { form: form })
+  res.render(Paths.eligibilityDefendantAgePage.associatedView, { form: form })
 }
 
 /* tslint:disable:no-default-export */
 export default express.Router()
-  .get(Paths.eligibilityDefendantAgeOptionPage.uri, (req: express.Request, res: express.Response): void => {
+  .get(Paths.eligibilityDefendantAgePage.uri, (req: express.Request, res: express.Response): void => {
     const draft: Draft<DraftClaim> = res.locals.claimDraft
     renderView(new Form(draft.document.eligibility), res)
   })
   .post(
-    Paths.eligibilityDefendantAgeOptionPage.uri,
+    Paths.eligibilityDefendantAgePage.uri,
     FormValidator.requestHandler(undefined, Eligibility.fromObject, ValidationGroups.OVER_18_DEFENDANT),
     ErrorHandling.apply(async (req: express.Request, res: express.Response, next: express.NextFunction): Promise<void> => {
       const form: Form<Eligibility> = req.body
