+++ src/main/views/macro/externalLink.njk
@@ -1,0 +1,3 @@
+{% macro externalLink(text, href) %}
+  <a href="{{ href }}" target="_blank">{{ text }}</a>
+{% endmacro %}
